refactor: split DirectTerminalHandler into separate file from DirectSessionHandler
- Moved terminal/shell handling to direct_terminal.py
- DirectSessionHandler now handles exec only
- shell_requested() returns DirectTerminalHandler instance for terminal sessions
- Added error handling in direct_terminal data_received and output_cb
