+++ api/tacticalrmm/tacticalrmm/gateway/handlers.py
@@ -8,12 +8,10 @@ from django.utils import timezone as djangotime
 from .audit import (
     _audit_exec_command,
     _audit_session_failed,
-    _audit_terminal_command,
     _close_session_and_audit,
-    _record_session_and_audit,
 )
+from .direct_terminal import DirectTerminalHandler
 from .exec import CmdProxy
-from .terminal import TerminalProxy
 
 logger = logging.getLogger("trmm")
 
@@ -61,13 +59,9 @@ class DirectSessionHandler(asyncssh.SSHServerSession):
         self._gateway_terminal_enabled = gateway_terminal_enabled
         self._gateway_session_timeout = gateway_session_timeout
         self._gateway_max_sessions = gateway_max_sessions
-        self._term = None
         self._exec = None
         self._chan = None
         self._started_at = None
-        self._terminal_type = ""
-        self._terminal_rows = 0
-        self._terminal_cols = 0
         self._session_type = None
         self._exec_cmd = None
         self._exec_completed = False
@@ -153,111 +147,15 @@ class DirectSessionHandler(asyncssh.SSHServerSession):
                 "SSH gateway terminal access is disabled by your administrator.\r\n",
                 audit_coro=audit,
             )
-        self._session_type = "shell"
-        shell = self._agent.effective_default_shell
-        self._term = TerminalProxy(self._agent, self._session_id, shell)
-        role_name = "None"
-        if self._user.role:
-            role_name = self._user.role.name
-        os_info = self._agent.operating_system or "Unknown"
-        agent_ver = self._agent.version or "Unknown"
-        pubip = self._agent.public_ip or "N/A"
-        local_ips = self._agent.local_ips() or "N/A"
-        self._chan.write(
-            f"\r\n\x1b[32mWelcome, \x1b[1m{self._user.username}\x1b[0m\x1b[32m [\x1b[1mRole: {role_name}\x1b[0m\x1b[32m]\x1b[0m\r\n"
-            f"\x1b[32mSSH session open on \x1b[1m{self._agent.hostname}\x1b[0m\x1b[32m "
-            f"({os_info}), shell: {shell}, agent: {agent_ver}, via TRMM proxy\x1b[0m\r\n"
-            f"\x1b[32mpubip: {pubip}, local IPs: {local_ips}\x1b[0m\r\n\r\n"
+        return DirectTerminalHandler(
+            self._user, self._agent, self._session_id, self._remote_ip,
+            client_version=self._client_version,
+            ssh_key_name=self._ssh_key_name,
+            ssh_key_type=self._ssh_key_type,
+            ssh_key_fingerprint=self._ssh_key_fingerprint,
+            gateway_session_timeout=self._gateway_session_timeout,
+            gateway_max_sessions=self._gateway_max_sessions,
         )
-        asyncio.create_task(self._start_terminal())
-        asyncio.create_task(
-            _record_session_and_audit(
-                self._user, self._agent, self._session_id, self._remote_ip,
-                client_version=self._client_version,
-                ssh_key_name=self._ssh_key_name,
-                ssh_key_type=self._ssh_key_type,
-                ssh_key_fingerprint=self._ssh_key_fingerprint,
-            )
-        )
-        logger.info(
-            "Gateway shell started user=%s agent=%s remote_ip=%s client=%s",
-            self._user.username, self._agent.agent_id, self._remote_ip, self._client_version,
-        )
-        return True
-
-    async def _start_terminal(self):
-        async def output_cb(data, done=False, exit_code=None):
-            try:
-                if isinstance(data, bytes):
-                    data = data.decode("utf-8", errors="replace")
-                data = _strip_ansi(data)
-                if data and self._chan and not self._chan.is_closing():
-                    self._chan.write(data)
-                if done:
-                    asyncio.create_task(_close_session_and_audit(
-                        self._user, self._agent, self._session_id, self._remote_ip,
-                        self._started_at,
-                        terminal_type=self._terminal_type,
-                        terminal_rows=self._terminal_rows,
-                        terminal_cols=self._terminal_cols,
-                    ))
-                    self._chan.exit(exit_code or 0)
-            except Exception:
-                logger.error("Gateway terminal output_cb error", exc_info=True)
-
-        try:
-            await self._term.start(output_cb)
-        except Exception as e:
-            logger.error("Gateway terminal start failed: %s", e, exc_info=True)
-            self._chan.write(f"\r\nFailed to start terminal: {e}\r\n")
-            self._chan.exit(1)
-
-    def pty_requested(self, term_type, term_size, term_modes):
-        if not self._gateway_terminal_enabled:
-            return False
-        self._terminal_type = term_type
-        if term_size:
-            self._terminal_cols, self._terminal_rows = term_size[0], term_size[1]
-        return True
-
-    def terminal_modes(self):
-        return {
-            asyncssh.VEOF: 4,
-            asyncssh.VINTR: 3,
-            asyncssh.VKILL: 21,
-            asyncssh.VQUIT: 28,
-            asyncssh.VSTART: 17,
-            asyncssh.VSTOP: 19,
-            asyncssh.VSUSP: 26,
-            asyncssh.VTIME: 0,
-            asyncssh.VMIN: 1,
-            asyncssh.ECHO: 0,
-            asyncssh.ECHOE: 0,
-            asyncssh.ECHOK: 0,
-            asyncssh.ECHOKE: 0,
-            asyncssh.ECHOCTL: 0,
-            asyncssh.ECHOPRT: 0,
-            asyncssh.ISIG: 1,
-            asyncssh.ICANON: 1,
-            asyncssh.IEXTEN: 1,
-            asyncssh.CTERMINAL: 0,
-        }
-
-    def data_received(self, data, datatype):
-        if self._term:
-            self._term_buf = getattr(self, "_term_buf", "") + data
-            if b"\r" in self._term_buf or b"\n" in self._term_buf:
-                lines = self._term_buf.replace(b"\r\n", b"\n").replace(b"\r", b"\n").split(b"\n")
-                self._term_buf = lines.pop()
-                for line in lines:
-                    if line.strip() and not any(b < 32 for b in line if b not in (9,)):
-                        asyncio.create_task(_audit_terminal_command(
-                            self._user, self._agent, line.decode("utf-8", errors="replace"),
-                        ))
-            try:
-                asyncio.create_task(self._term.write(data))
-            except Exception as e:
-                logger.error("Gateway terminal write failed: %s", e)
 
     def eof_received(self):
         if self._session_type == "exec" and not self._exec_completed:
@@ -267,34 +165,18 @@ class DirectSessionHandler(asyncssh.SSHServerSession):
     def connection_lost(self, exc):
         if self._session_type == "exec" and not self._exec_completed:
             return
-        if self._term:
-            asyncio.create_task(self._term.stop())
         if self._exec:
             asyncio.create_task(self._exec.stop())
         if exc:
             logger.error("Gateway connection lost: %s", exc)
 
-    def terminal_size_changed(self, w, h, pw, ph):
-        if self._term:
-            asyncio.create_task(self._term.resize(h, w))
-
     def closed(self):
-        if self._term:
-            asyncio.create_task(self._term.stop())
         if self._exec:
             asyncio.create_task(self._exec.stop())
-        if self._started_at and self._session_type == "shell":
+        if self._started_at and self._session_type == "exec":
             asyncio.create_task(
                 _close_session_and_audit(
                     self._user, self._agent, self._session_id, self._remote_ip,
                     self._started_at,
-                    terminal_type=self._terminal_type,
-                    terminal_rows=self._terminal_rows,
-                    terminal_cols=self._terminal_cols,
                 )
-            )
-            logger.info(
-                "Gateway session ended user=%s agent=%s duration=%ds",
-                self._user.username, self._agent.agent_id,
-                int((djangotime.now() - self._started_at).total_seconds()),
             )
